# src/Resources/schema/quote.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "com.shopware.quote",
    "version": "2026-04-08",
    "description": "Buyer-facing B2B Request-for-Quote capability for Shopware shops. Protocol-neutral: any UCP buyer agent can use it.\n\n## Authentication and authorization\n\nQuote operations act on behalf of a customer, so they require a **customer-linked context token** sent as `sw-context-token`, obtained through the shop's `dev.ucp.common.identity_linking` (OAuth 2.0) flow. The token is the trust boundary: the customer's consent - not anything in the request body - decides whose quotes are touched, and agents never hold customer credentials. Requests still carry the standard `UCP-Agent` header (and a request signature when the shop's signature policy requires it), exactly like every other UCP runtime route.\n\nThe customer must additionally have the commercial Quote Management feature enabled (`customer_specific_features`). Failures use the standard UCP error envelope:\n\n| Check | Result |\n|---|---|\n| Capability not enabled for the sales channel, or SwagCommercial / Quote Management not licensed | unsupported capability (the capability is then also absent from /.well-known/ucp) |\n| Context token missing or not linked to a customer | validation error naming `$.headers.sw-context-token` |\n| Customer lacks the Quote Management feature | unsupported capability |\n| Quote id unknown or owned by another customer | not found (existence is never confirmed) |\n| Accepting an expired or non-actionable quote | error from the commercial quote-to-order flow |\n\n## State machine\n\nBuyer-visible states and who acts (full transition table in `x-state-machine`):\n\n- `open` - request submitted, merchant's turn. Poll.\n- `in_review` - merchant is processing. Poll.\n- `replied` - merchant made an offer, **buyer's turn**: accept, counter, or decline.\n- `change_requested` - buyer countered, merchant's turn. Poll.\n- `accepted`, `declined`, `expired`, `withdrawn`, `cancelled` - terminal for the buyer.\n\n## Expiration\n\nOffers carry `expiration_date` (ISO 8601, always present in responses, may be null before the merchant replies). A quote can be auto-flagged `expired` by a background job; accepting an expired quote fails. Agents must read `expiration_date` from the offer and act before it.\n\n## Price semantics\n\nAll amounts are in the quote's `currency`. `totals.tax_status` states whether `totals.gross` (tax included) or `totals.net` is authoritative. Line item `unit_price`, `requested_unit_price` and counter-offer `requested_unit_price` are **per unit**. Requested prices are the buyer's ask; the merchant's offer is reflected in `unit_price`/`totals` once the quote reaches `replied`.\n\n## Async negotiation\n\nThe merchant side (possibly automated) responds out-of-band. Agents poll `GET /ucp/quotes/{id}` for state changes; recommended polling interval: 300 seconds (see `x-polling-interval-seconds`). Webhooks are not part of this capability version.",
    "x-capability": "com.shopware.quote"
  },
  "x-polling-interval-seconds": 300,
  "x-state-machine": {
    "states": {
      "open": {
        "actor": "merchant",
        "buyer_actions": []
      },
      "in_review": {
        "actor": "merchant",
        "buyer_actions": []
      },
      "replied": {
        "actor": "buyer",
        "buyer_actions": [
          "accept",
          "counter",
          "decline"
        ]
      },
      "change_requested": {
        "actor": "merchant",
        "buyer_actions": []
      },
      "accepted": {
        "actor": "none",
        "terminal": true,
        "buyer_actions": []
      },
      "declined": {
        "actor": "none",
        "terminal": true,
        "buyer_actions": []
      },
      "expired": {
        "actor": "merchant",
        "terminal": true,
        "buyer_actions": [],
        "note": "merchant may extend expiration back to replied"
      },
      "withdrawn": {
        "actor": "none",
        "terminal": true,
        "buyer_actions": []
      },
      "cancelled": {
        "actor": "none",
        "terminal": true,
        "buyer_actions": []
      }
    },
    "transitions": [
      {
        "from": "draft",
        "action": "customer_send",
        "to": "open",
        "actor": "buyer",
        "note": "performed automatically inside POST /ucp/quotes; draft is never exposed"
      },
      {
        "from": "open",
        "action": "process",
        "to": "in_review",
        "actor": "merchant"
      },
      {
        "from": "open",
        "action": "sent",
        "to": "replied",
        "actor": "merchant"
      },
      {
        "from": "in_review",
        "action": "sent",
        "to": "replied",
        "actor": "merchant"
      },
      {
        "from": "replied",
        "action": "accept",
        "to": "accepted",
        "actor": "buyer",
        "endpoint": "POST /ucp/quotes/{id}/accept"
      },
      {
        "from": "replied",
        "action": "decline",
        "to": "declined",
        "actor": "buyer",
        "endpoint": "POST /ucp/quotes/{id}/decline"
      },
      {
        "from": "replied",
        "action": "request_change",
        "to": "change_requested",
        "actor": "buyer",
        "endpoint": "POST /ucp/quotes/{id}/counter"
      },
      {
        "from": "replied",
        "action": "expire",
        "to": "expired",
        "actor": "system"
      },
      {
        "from": "change_requested",
        "action": "admin_resend",
        "to": "replied",
        "actor": "merchant"
      },
      {
        "from": "expired",
        "action": "admin_extend_expiration",
        "to": "replied",
        "actor": "merchant"
      },
      {
        "from": "declined",
        "action": "reopen",
        "to": "reopen",
        "actor": "merchant"
      }
    ]
  },
  "security": [
    {
      "contextToken": [],
      "ucpAgent": []
    }
  ],
  "paths": {
    "/ucp/quotes": {
      "post": {
        "operationId": "requestQuote",
        "summary": "Request a quote (RFQ). Creates and submits the request in one call; the returned quote is in state `open`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote created and submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "get": {
        "operationId": "listQuotes",
        "summary": "The customer's quotes, newest first. Only quotes owned by the customer the credential resolves to are listed. Use this to rediscover a quote whose id was not stored, or to poll several negotiations at once.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 25
            },
            "description": "Page size; values above 50 are clamped."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteList"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/ucp/quotes/{id}": {
      "get": {
        "operationId": "getQuote",
        "summary": "Read a quote snapshot (state, line items with offered prices, totals, expiration_date, comments). Only quotes owned by the resolved customer are visible; anything else is 404.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          }
        ],
        "responses": {
          "200": {
            "description": "Quote snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/ucp/quotes/{id}/counter": {
      "post": {
        "operationId": "counterQuote",
        "summary": "Counter-offer: new per-line requested unit prices and/or a comment. Valid only in state `replied`; transitions to `change_requested`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated quote (state change_requested)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/ucp/quotes/{id}/accept": {
      "post": {
        "operationId": "acceptQuote",
        "summary": "Accept the offer. Accepting IS ordering: executes the quote-to-order flow and returns the order reference. Valid only in state `replied` and before expiration_date.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Not acceptable in current state or expired (code CHECKOUT__QUOTE_CANNOT_PLACE_ORDER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/ucp/quotes/{id}/decline": {
      "post": {
        "operationId": "declineQuote",
        "summary": "Decline the offer with an optional comment. Valid only in state `replied`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/quoteId"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Declined quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "contextToken": {
        "type": "apiKey",
        "in": "header",
        "name": "sw-context-token",
        "description": "Customer-linked Shopware context token obtained through the shop's dev.ucp.common.identity_linking (OAuth 2.0) flow. Identifies the customer the agent acts for."
      },
      "ucpAgent": {
        "type": "apiKey",
        "in": "header",
        "name": "UCP-Agent",
        "description": "Standard UCP agent identification header (profile URL). Signature requirements follow the sales channel signature policy."
      }
    },
    "parameters": {
      "quoteId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-f]{32}$"
        },
        "description": "Quote id (UUID hex). Ids are unguessable but treated as secrets of the owning customer: foreign ids return 404."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid payload or state",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Capability disabled, not licensed, or Quote Management not enabled for the customer",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Quote unknown or owned by another customer",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "QuoteRequest": {
        "type": "object",
        "required": [
          "line_items"
        ],
        "properties": {
          "line_items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "description": "product_id or product_number is required",
              "properties": {
                "product_id": {
                  "type": "string"
                },
                "product_number": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1
                },
                "requested_unit_price": {
                  "type": "number",
                  "description": "Buyer's asking price per unit, in the shop's customer-facing currency"
                }
              },
              "required": [
                "quantity"
              ]
            }
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "CounterRequest": {
        "type": "object",
        "properties": {
          "line_items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Reference an existing quote line by id (preferred) or product_id",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Quote line item id from GET /ucp/quotes/{id}"
                },
                "product_id": {
                  "type": "string"
                },
                "requested_unit_price": {
                  "type": "number",
                  "description": "New asking price per unit"
                }
              },
              "required": [
                "requested_unit_price"
              ]
            }
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "DeclineRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        }
      },
      "Quote": {
        "type": "object",
        "required": [
          "id",
          "state",
          "expiration_date",
          "totals",
          "line_items"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "quote_number": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "in_review",
              "replied",
              "change_requested",
              "accepted",
              "declined",
              "expired",
              "withdrawn",
              "cancelled",
              "reopen"
            ]
          },
          "expiration_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Always present. Accept before this instant; null means the merchant has not set one yet (the quote may still be expired by a background job)."
          },
          "currency": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 4217 code; all amounts in this document are in this currency"
          },
          "totals": {
            "type": "object",
            "properties": {
              "gross": {
                "type": [
                  "number",
                  "null"
                ],
                "description": "Total including tax"
              },
              "net": {
                "type": [
                  "number",
                  "null"
                ],
                "description": "Total excluding tax"
              },
              "tax_status": {
                "type": "string",
                "enum": [
                  "gross",
                  "net",
                  "tax-free"
                ],
                "description": "Which total is the customer-facing authoritative amount"
              }
            }
          },
          "line_items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "product_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unit_price": {
                  "type": "number",
                  "description": "Current (offered) price per unit"
                },
                "total_price": {
                  "type": "number"
                },
                "requested_unit_price": {
                  "type": [
                    "number",
                    "null"
                  ],
                  "description": "Buyer's last ask per unit, if any"
                }
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string"
                },
                "author": {
                  "type": "string",
                  "enum": [
                    "buyer",
                    "merchant"
                  ]
                },
                "created_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                }
              }
            }
          },
          "order_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set after acceptance"
          },
          "ucp": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string"
              }
            }
          }
        }
      },
      "AcceptResponse": {
        "type": "object",
        "required": [
          "quote_id",
          "state",
          "order"
        ],
        "properties": {
          "quote_id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "const": "accepted"
          },
          "order": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "order_number": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          "ucp": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string"
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "error"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "code": {
                  "type": "string",
                  "description": "Machine-readable error code",
                  "examples": [
                    "validation_error",
                    "resource_not_found",
                    "unsupported_capability"
                  ]
                },
                "content": {
                  "type": "string"
                },
                "severity": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Standard UCP error envelope produced by the SDK (validation, not-found, and unsupported-capability errors)."
      },
      "QuoteList": {
        "type": "object",
        "required": [
          "quotes",
          "total",
          "limit",
          "page"
        ],
        "properties": {
          "quotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quote"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total quotes for this customer, not just this page"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        }
      }
    }
  }
}